Moved the variable pendulum to sphinx gallery.
